requirement.txt
dj-database-url
psycopg2-binary
django
djangorestframework
mysqlclient
gunicorn
django-cors-headers
djangorestframework-simplejwt
